+++ system/src/Grav/Framework/Flex/Traits/FlexMediaTrait.php
@@ -49,7 +49,7 @@ trait FlexMediaTrait
      */
     public function getStorageFolder()
     {
-        return $this->getFlexDirectory()->getStorageFolder($this->getStorageKey());
+        return $this->exists() ? $this->getFlexDirectory()->getStorageFolder($this->getStorageKey()) : '';
     }
 
     /**
@@ -57,7 +57,7 @@ trait FlexMediaTrait
      */
     public function getMediaFolder()
     {
-        return $this->getFlexDirectory()->getMediaFolder($this->getStorageKey());
+        return $this->exists() ? $this->getFlexDirectory()->getMediaFolder($this->getStorageKey()) : '';
     }
 
     /**
@@ -153,6 +153,12 @@ trait FlexMediaTrait
         /** @var UniformResourceLocator $locator */
         $locator = $grav['locator'];
         $path = $media->getPath();
+        if (!$path) {
+            $language = $grav['language'];
+
+            throw new RuntimeException($language->translate('PLUGIN_ADMIN.FAILED_TO_MOVE_UPLOADED_FILE'), 400);
+        }
+
         if ($locator->isStream($path)) {
             $path = $locator->findResource($path, true, true);
             $locator->clearCache($path);
@@ -202,12 +208,16 @@ trait FlexMediaTrait
         }
 
         $media = $this->getMedia();
+        $path = $media->getPath();
+        if (!$path) {
+            return;
+        }
 
         /** @var UniformResourceLocator $locator */
         $locator = $grav['locator'];
 
-        $targetPath = $media->getPath() . '/' . $dirname;
-        $targetFile = $media->getPath() . '/' . $filename;
+        $targetPath = $path . '/' . $dirname;
+        $targetFile = $path . '/' . $filename;
         if ($locator->isStream($targetFile)) {
             $targetPath = $locator->findResource($targetPath, true, true);
             $targetFile = $locator->findResource($targetFile, true, true);